@@ -8,7 +8,7 @@ from ezrules.models.database import Base, engine
 from ezrules.models.history_meta import versioned_session
 
 
-@pytest.fixture(scope="session")
+@pytest.fixture(scope="function")
 def logged_out_manager_client():
     ezruleapp.app.config["TESTING"] = True
     ezruleapp.app.config["WTF_CSRF_METHODS"] = []
@@ -17,7 +17,7 @@ def logged_out_manager_client():
         yield client
 
 
-@pytest.fixture(scope="session")
+@pytest.fixture(scope="function")
 def logged_out_eval_client():
     ezrulevalapp.app.config["TESTING"] = True
 
@@ -31,14 +31,19 @@ def engine_fix():
         drop_database(engine.url)
     create_database(engine.url)
 
-    Base.metadata.create_all(
-        engine
-    )  # Assuming Base is the declarative base from your models
+    Base.metadata.create_all(engine)  # Assuming Base is the declarative base from your models
 
     yield engine
 
-    Base.metadata.drop_all(engine)
-    drop_database(engine.url)
+    # Close all connections before dropping database
+    engine.dispose()
+
+    # Try to drop the database, ignore if still in use
+    try:
+        drop_database(engine.url)
+    except Exception:
+        # Database might still be in use, which is okay for tests
+        pass
 
 
 @pytest.fixture(scope="session")
@@ -59,8 +64,8 @@ def session(connection):
     session.add(org)
     session.commit()
 
-    admin_email = f"admin@test_org.com"
-    admin_password = f"12345678"
+    admin_email = "admin@test_org.com"
+    admin_password = "12345678"
     session.add(
         User(
             email=admin_email,
@@ -75,16 +80,47 @@ def session(connection):
     ezruleapp.fsrm.o_id = org.o_id
     ezruleapp.rule_engine_config_producer.db = session
     ezruleapp.rule_engine_config_producer.o_id = org.o_id
+    ezruleapp.outcome_manager.db_session = session
+    ezruleapp.outcome_manager.o_id = org.o_id
+    ezruleapp.outcome_manager._initialized = False
+    ezruleapp.outcome_manager._cached_outcomes = None
+    ezruleapp.user_list_manager.db_session = session
+    ezruleapp.user_list_manager.o_id = org.o_id
+    ezruleapp.user_list_manager._initialized = False
+    ezruleapp.user_list_manager._cached_lists = None
+
+    # Set up application context for tests
+    from ezrules.core.application_context import reset_context, set_organization_id, set_user_list_manager
+    from ezrules.core.user_lists import PersistentUserListManager
+
+    reset_context()  # Reset context between tests
+    test_list_provider = PersistentUserListManager(db_session=session, o_id=org.o_id)
+    set_organization_id(org.o_id)
+    set_user_list_manager(test_list_provider)
     ezrulevalapp.lre.db = session
     ezrulevalapp.lre.o_id = org.o_id
 
     yield session
 
+    # Clean up Flask app references to avoid lingering connections
+    ezruleapp.fsrm.db = None
+    ezruleapp.rule_engine_config_producer.db = None
+    ezruleapp.outcome_manager.db_session = None
+    ezruleapp.outcome_manager._initialized = False
+    ezruleapp.outcome_manager._cached_outcomes = None
+    ezruleapp.user_list_manager.db_session = None
+    ezruleapp.user_list_manager._initialized = False
+    ezruleapp.user_list_manager._cached_lists = None
+
+    # Clean up application context
+    reset_context()
+    ezrulevalapp.lre.db = None
+
     session.close()
     transaction.rollback()
 
 
-@pytest.fixture
+@pytest.fixture(scope="function")
 def logged_in_manager_client(session, logged_out_manager_client):
     # Log in the test user
     logged_out_manager_client.post(
