@@ -1,13 +1,12 @@
-from typing import Dict
-
 from pydantic import BaseModel, field_validator
+
 from ezrules.models.backend_core import TestingRecordLog, TestingResultsLog
 
 
 class Event(BaseModel):
     event_id: str
     event_timestamp: int  # Assuming Unix timestamp as input
-    event_data: Dict
+    event_data: dict
 
     @field_validator("event_timestamp", mode="before")
     def validate_unix_timestamp(cls, value):
@@ -17,9 +16,7 @@ class Event(BaseModel):
 
         # Ensure the timestamp is in a reasonable range (e.g., 1970-01-01 to 3000-01-01)
         min_timestamp = 0  # Unix timestamp for 1970-01-01T00:00:00Z
-        max_timestamp = (
-            32503680000  # Approximate Unix timestamp for 3000-01-01T00:00:00Z
-        )
+        max_timestamp = 32503680000  # Approximate Unix timestamp for 3000-01-01T00:00:00Z
         if not (min_timestamp <= value <= max_timestamp):
             raise ValueError("Timestamp out of range")
 
