@@ -1,4 +1,4 @@
-from flask import Flask, request, abort, jsonify
+from flask import Flask, abort, jsonify, request
 from pydantic import ValidationError
 
 from ezrules.backend.data_utils import Event, eval_and_store
@@ -16,8 +16,12 @@ lre = LocalRuleExecutorSQL(db=db_session, o_id=o_id)
 def evaluate():
     request_data = request.get_json()
     try:
-        event = Event(**request_data)
-    except ValidationError as e:
+        event = Event(
+            event_id=request_data["event_id"],
+            event_timestamp=request_data["event_timestamp"],
+            event_data=request_data["event_data"],
+        )
+    except ValidationError:
         abort(400, description="Bad Request: Could not validate the json structure")
     response = eval_and_store(lre, event)
     return jsonify(response)
@@ -25,4 +29,4 @@ def evaluate():
 
 @app.route("/ping", methods=["GET"])
 def ping():
-    return f"OK"
+    return "OK"
