@@ -1,7 +1,9 @@
 import ast
 import json
-import pyparsing as pp
 from typing import Any
+
+import pyparsing as pp
+
 from ezrules.core.user_lists import AbstractUserListManager
 
 
@@ -24,9 +26,7 @@ class DollarNotationConverter:
         return f't["{tokens[0][1:]}"]'
 
     def __init__(self):
-        search_for_word = pp.Combine(
-            self.TRIGGER_CHAR + pp.Word(pp.alphas + "_", pp.alphanums + "_")
-        )
+        search_for_word = pp.Combine(self.TRIGGER_CHAR + pp.Word(pp.alphas + "_", pp.alphanums + "_"))
 
         line_parser = search_for_word
         line_parser.ignore(pp.QuotedString('"'))
