@@ -1,6 +1,6 @@
 import pytest
 from pydantic import ValidationError
-from typing import Dict
+
 from ezrules.backend.data_utils import Event
 
 
@@ -36,12 +36,10 @@ from ezrules.backend.data_utils import Event
         ),
     ],
 )
-def test_event(event_id: str, event_timestamp: int, event_data: Dict, expected: bool):
+def test_event(event_id: str, event_timestamp: int, event_data: dict, expected: bool):
     if expected:
         # If expected is True, we expect the Event to be created successfully
-        event = Event(
-            event_id=event_id, event_timestamp=event_timestamp, event_data=event_data
-        )
+        event = Event(event_id=event_id, event_timestamp=event_timestamp, event_data=event_data)
         assert event.event_id == event_id
         assert event.event_timestamp == event_timestamp
         assert event.event_data == event_data
